Create exec, create lib, move module to own repo
