+++ lib/rpc.go
@@ -6,21 +6,9 @@ import (
 	"net"
 	"net/rpc"
 	"net/rpc/jsonrpc"
-	"strconv"
 )
 
-func getOpenPort() string {
-	for i := 65535; i >= 49152; i-- {
-		conn, err := net.Listen("tcp", ":"+strconv.Itoa(i))
-		if err == nil {
-			conn.Close()
-			return strconv.Itoa(i)
-		}
-	}
-	return ""
-}
-
-func rpcCodecClient() rpc.ClientCodec {
+func RpcCodecClient() rpc.ClientCodec {
 	conn, err := net.Dial("tcp", "localhost:5555")
 	if err != nil {
 		fmt.Println(err)
@@ -29,7 +17,7 @@ func rpcCodecClient() rpc.ClientCodec {
 	return jsonrpc.NewClientCodec(conn)
 }
 
-func rpcCodecClientWithPort(port string) rpc.ClientCodec {
+func RpcCodecClientWithPort(port string) rpc.ClientCodec {
 	conn, err := net.Dial("tcp", "localhost:"+port)
 	if err != nil {
 		fmt.Println(err)
@@ -38,6 +26,6 @@ func rpcCodecClientWithPort(port string) rpc.ClientCodec {
 	return jsonrpc.NewClientCodec(conn)
 }
 
-func rpcCodecServer(conn io.ReadWriteCloser) rpc.ServerCodec {
+func RpcCodecServer(conn io.ReadWriteCloser) rpc.ServerCodec {
 	return jsonrpc.NewServerCodec(conn)
 }
